[Rebuild] Post, User uuid 타입의 identifier를 String -> YeogaTypealias에 지정된 타입들로 전부 리빌딩했습니다.

+++ travelPlan/Source/Presentation/Post/Common/PostNotificationManager.swift
@@ -21,7 +21,7 @@ final class PostNotificationManager {
 
 // MARK: - For post option
 extension PostNotificationManager {
-  func notifyPostHasBlocked(postId: Int32, postOptionLocation: PostOptionLocation) {
+  func notifyPostHasBlocked(postId: Int64, postOptionLocation: PostOptionLocation) {
     NotificationCenter.default.post(
       name: .hasPostBlocked,
       object: nil,
@@ -29,7 +29,7 @@ extension PostNotificationManager {
                  "postOptionLocation": postOptionLocation])
   }
   
-  func notifyUserWantToSharePost(postId: Int32, postTitle: String) {
+  func notifyUserWantToSharePost(postId: Int64, postTitle: String) {
     NotificationCenter.default.post(
       name: .postShareFromPostOptionActionSheet,
       object: nil,
